requirements.txt
beautifulsoup4==4.6.1
boardgamegeek2==0.0.1
bs4==0.0.1
certifi==2018.4.16
chardet==3.0.4
docutils==0.14
idna==2.7
Kivy==1.10.1
Kivy-Garden==0.1.4
kivy.deps.glew==0.1.10
kivy.deps.sdl2==0.1.18
Pygments==2.2.0
pypiwin32==223
pywin32==223
requests==2.19.1
requests-cache==0.4.13
